changement des valeurs des id de int à double pour le DAO

+++ src/Models/CourseType.java
@@ -1,21 +1,21 @@
 package Models;
 
 public class CourseType {
-    private int ID;
+    private Long ID;
     private enum Type {INTERACTIF,MAGISTRAL, TD,TP,PROJET,SOUTIEN}
 
     public Type type;
 
-    public CourseType(int ID, Type type) {
+    public CourseType(Long ID, Type type) {
         this.ID = ID;
         this.type = type;
     }
 
-    public int getID() {
+    public Long getID() {
         return ID;
     }
 
-    public void setID(int ID) {
+    public void setID(Long ID) {
         this.ID = ID;
     }
